# server/common/api.yml
openapi: 3.0.1
info:
  title: Picard Datahub Api
  version: 0.5.2
  description: >-
    The Picard Datahub Api is the public interface between the Datahub and
    Picard
  contact:
    name: Friedrich PICARD GmbH & Co.KG
    url: 'https://www.picard.de'
    email: edv@picard.de
  x-logo:
    url: ''
servers:
  - url: 'https://datahub.picard.de:3500/'
  - url: 'http://localhost:3500/'
paths:
...
  /api/v1/datahub/manufacturers/:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductManufacturer'
          description: Responses with a list of product categories
          x-oad-type: response
      operationId: Receive Manufacturer List
      summary: Get all product manufacturers
      description: This Routes sends you all available Manufacturers
    x-oad-type: operation
  '/api/v1/datahub/downloads/files/id/{fileid}':
    get:
      tags:
        - Files
      parameters:
        - name: fileid
          description: The associated File Id to this request
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                format: binary
          description: Responds with the associated File
          x-oad-type: response
      operationId: Receive Specific File by id
      summary: Get a specific uploaded file
      description: Download Files with this Route
    x-oad-type: operation
  '/api/v1/datahub/downloads/files/name/{filename}':
    get:
      tags:
        - Files
      parameters:
        - name: filename
          description: The associated File Id to this request
          schema:
            type: string
          in: path
          required: true
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                format: binary
          description: Responds with the associated File
          x-oad-type: response
      operationId: Receive Specific File by originalfilename
      summary: Get a specific uploaded file
      description: Download Files with this Route
    x-oad-type: operation
  /api/v1/datahub/uploads/files/:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  nullable: false
                  format: binary
                  type: string
      tags:
        - Files
      parameters:
        - name: filetype
          schema:
            nullable: false
            enum:
              - LIEFERSCHEIN
              - MEDIA
              - DOKUMENT
              - IMAGE
              - VIDEO
              - ANY
            type: string
          in: query
        - name: additionaldata
          schema:
            type: object
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
          description: Responds with status of success
          x-oad-type: response
      operationId: Transmit a file to api
      summary: Stores a  file
      description: This Routes is used to upload files
    x-oad-type: operation
...
  /api/v1/datahub/orders/:
    summary: PDO Orders to Picard
    post:
      requestBody:
        description: The complete Order Object, See example or Schema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
        required: true
      tags:
        - Work in Progress
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  title: ProductData
                  type: object
                  x-oad-type: object
          description: Responses with an object containing all changes
          x-oad-type: response
      operationId: Post Orders from PDO Portal
      summary: Send PDO Orders to Picard
      description: Transfer Route to transfer orders to Picard
    x-oad-type: operation
    x-last-modified: 1598348347635
  /api/v1/datahub/pdo/settings:
    summary: PDO Settings from Picard
    get:
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      tags:
        - Work in Progress
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PDOSettings'
                  
          description: Responses with an object containing all settings
          x-oad-type: response
      operationId: Get PDO Settings from Picard
      summary: Get PDO Settings from Picard
      description: Transfer Route to transfer settings from Picard
    x-oad-type: operation
    x-last-modified: 1598348347635
components:
  schemas:
...
    ProductManufacturer:
      type: object
      properties:
        ManufacturerId:
          type: number
        FullName:
          type: string
        ShortName:
          type: string
        CreatedAt:
          type: string
          format: date-time
        ModifiedAt:
          type: string
          format: date-time
        DeletedAt:
          type: string
          format: date-time
...
    PDOSettings:
      type: object
      properties:
        CustomerId:
          type: number
        CustomerName:
          type: string
        CustomerKey:
          type: string
        PriceProfile:
          type: number
        IsSelection:
          type: boolean
        IsOccasion:
          type: boolean
        Localization:
          type: string
        CreatedAt:
          type: string
          format: date-time
        ModifiedAt:
          type: string
          format: date-time
        DeletedAt:
          type: string
          format: date-time
...
    OrderRequest:
      type: object
      properties:
        CreatedAt:
          type: string
          format: date-time
        OrderId:
          type: number
        InvoiceNumber:
          type: number
        InvoiceDate:
          type: string
          format: date-time
        DeliveryDate:
          type: string
          format: date-time
        PicardCustomerId:
          type: number
        ShopCustomerId:
          type: number
        ShopLoginId:
          type: number
        ShopLoginEmail:
          example: 'name@domain'
          type: string
        AddressDelivery:
          type: object
          properties:
            Country:
              example: 'Germany'
              type: string
            State:
              example: 'NRW'
              type: string
            City:
              example: 'Bochum'
              type: string
            Postcode:
              example: '44805'
              type: string
            Street:
              example: 'Dietrich-Benking-Straße 78'
              type: string
            Company:
              example: 'Friedrich PICARD GmbH & Co. KG'
              type: string
            Firstname:
              example: 'Max'
              type: string
            Lastname:
              example: 'Mustermann'
              type: string
            Phone:
              example: '+49 234 53993-130'
              type: string
        AddressInvoice:
          type: object
          properties:
            Country:
              example: 'Germany'
              type: string
            State:
              example: 'NRW'
              type: string
            City:
              example: 'Bochum'
              type: string
            Postcode:
              example: '44805'
              type: string
            Street:
              example: 'Dietrich-Benking-Straße 78'
              type: string
            Company:
              example: 'Friedrich PICARD GmbH & Co. KG'
              type: string
            Firstname:
              example: 'Max'
              type: string
            Lastname:
              example: 'Mustermann'
              type: string
            Phone:
              example: '+49 234 53993-130'
              type: string
        Currency:
          enum:
            - EUR
            - USD
            - GBP
          type: string
          minLength: 2
          maxLength: 3
        Language:
          enum:
            - DE
            - EN
          type: string
          minLength: 2
          maxLength: 2
        VatNumber:
          example: 'DE 286241645'
          type: string
        CarrierId:
          example: 'placeholder'
          type: string
        IsPDFDelivered:
          type: boolean
        DeliveryNote:
          example: 'placeholder'
          type: string
        IsPaid:
          type: boolean
        PaymentMethod:
          type: string
        TotalPaid:
          example: '0.00'
          type: string
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
        TotalShipping:
          example: '0.00'
          type: string
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
        OrderRows:
          type: array
          items:
            type: object
            properties:
              ProductId:
                type: number
              ProductReference:
                type: string
              ProductQuantity:
                example: 1
                multipleOf: 1
                type: number
              Price:
                example: '0.00'
                type: string
                pattern: '^[0-9]+(\.[0-9]{1,2})?$'
              ProductSupplierReference:
                example: 'placeholder'
                type: string
              LineReferenceText:
                example: 'placeholder'
                type: string
              LinearLength:
                example: '0.00'
                type: string
                pattern: '^[0-9]+(\.[0-9]{1,2})?$'
              LinearDistanceFirstBore:
                example: '0.00'
                type: string
                pattern: '^[0-9]+(\.[0-9]{1,2})?$'
              LinearSymmetrical:
                type: number
        ShopId:
          type: number
        Reference:
          example: 'placeholder'
          type: string
        Comment:
          example: 'placeholder'
          type: string
        DiscountName:
          example: 'placeholder'
          type: string
        DiscountCode:
          example: 'placeholder'
          type: string
        DiscountValue:
          example: '0.00'
          type: string
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
...
  securitySchemes:
    Development:
      type: apiKey
      description: This Api Key is for accessing developer routes
      name: token
      in: header
...
security:
  - Development: []
...